Get Alpine Docker build working.
Executable was crashing when parsing the reaction XML when the first app instance was loading.
Moving the reaction db initialization to happen before the first load fixed this.
I assume this is because musl has a smaller default stack size than glibc - but didnt deeply investigate.

Also changed a few other things around so the executable created for the docker build is static.
And a few other little things.